Renamed the test file.
